pyproject.toml
[project]
name = "context-windows-lab"
version = "1.0.0"
description = "Context Windows in Practice - LLM Lab Experiments"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Omer", email = "omer@example.com"},
    {name = "Yogev Cuperman", email = "yogev@example.com"}
]
keywords = ["llm", "context-windows", "rag", "langchain", "experiments"]

dependencies = [
    "anthropic>=0.39.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.3.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.13.0",
    "scipy>=1.11.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.7.0",
    "tqdm>=4.66.0",
    "rich>=13.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "reportlab>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[project.scripts]
context-lab = "context_windows_lab.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/context_windows_lab"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=src/context_windows_lab --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
